Refactor generateJobDefaults
