+++ src/tapis-api/utils/jobDefaults.ts
@@ -1,0 +1,57 @@
+import { Apps, Jobs, Systems } from '@tapis/tapis-typescript';
+import { generateRequiredFileInputsFromApp } from 'tapis-api/utils/jobFileInputs';
+import { generateRequiredFileInputArraysFromApp } from 'tapis-api/utils/jobFileInputArrays';
+import { generateJobArgsFromSpec } from 'tapis-api/utils/jobArgs';
+
+const generateJobDefaults = ({
+  app,
+  systems,
+}: {
+  app?: Apps.TapisApp;
+  systems: Array<Systems.TapisSystem>;
+}): Partial<Jobs.ReqSubmitJob> => {
+  if (!app) {
+    return {};
+  }
+  const systemDefaultQueue = systems.find(
+    (system) => system.id === app.jobAttributes?.execSystemId
+  )?.batchDefaultLogicalQueue;
+  const defaultValues: Partial<Jobs.ReqSubmitJob> = {
+    name: `${app.id}-${app.version}-${new Date().toISOString().slice(0, -5)}`,
+    description: app.description,
+    appId: app.id,
+    appVersion: app.version,
+    archiveOnAppError: app.jobAttributes?.archiveOnAppError ?? true,
+    archiveSystemId: app.jobAttributes?.archiveSystemId,
+    archiveSystemDir: app.jobAttributes?.archiveSystemDir,
+    nodeCount: app.jobAttributes?.nodeCount,
+    coresPerNode: app.jobAttributes?.coresPerNode,
+    jobType: app.jobType,
+    memoryMB: app.jobAttributes?.memoryMB,
+    maxMinutes: app.jobAttributes?.maxMinutes,
+    isMpi: app.jobAttributes?.isMpi,
+    mpiCmd: app.jobAttributes?.mpiCmd,
+    cmdPrefix: app.jobAttributes?.cmdPrefix,
+    execSystemId: app.jobAttributes?.execSystemId,
+    execSystemLogicalQueue:
+      app.jobAttributes?.execSystemLogicalQueue ?? systemDefaultQueue,
+    fileInputs: generateRequiredFileInputsFromApp(app),
+    fileInputArrays: generateRequiredFileInputArraysFromApp(app),
+    parameterSet: {
+      appArgs: generateJobArgsFromSpec(
+        app.jobAttributes?.parameterSet?.appArgs ?? []
+      ),
+      containerArgs: generateJobArgsFromSpec(
+        app.jobAttributes?.parameterSet?.containerArgs ?? []
+      ),
+      schedulerOptions: generateJobArgsFromSpec(
+        app.jobAttributes?.parameterSet?.schedulerOptions ?? []
+      ),
+      archiveFilter: app.jobAttributes?.parameterSet?.archiveFilter,
+      envVariables: app.jobAttributes?.parameterSet?.envVariables,
+    },
+  };
+  return defaultValues;
+};
+
+export default generateJobDefaults;
